requirements.txt
requests
pyyaml
pyiceberg
pyarrow
s3fs
sqlalchemy
boto3